Add missing type hints to preprocessor.py
Adding these type hints highlighted a few places where the code is quite
confusing, and I've added FIXME comments to remind us of those.

There was one block of code that relied on dynamic typing to such an extent
that I couldn't find a simple way to restructure it. I've added comments
there to disable mypy for now, but we should come back and rewrite that
function when we have a better idea of how it works.

Signed-off-by: John Pennycook <john.pennycook@intel.com>