Rename Platform class to Preprocessor class
This is a much better name for the class, because it represents the state of a
specific instance of a C preprocessor.

The rename also makes it easier to justify why things like include paths and
defines are an essential part of construction, allowing these to be moved into
the __init__() function.

Signed-off-by: John Pennycook <john.pennycook@intel.com>