Simplify some Preprocessor member functions
define() previously took a name and a macro, despite a macro's name being
encoded in the macro itself.

undefine() and get_macro() expected a string, despite the rest of the
preprocessor code using an Identifier token. Using the Identifier directly
is less error-prone than having to remember to extract the string.

Signed-off-by: John Pennycook <john.pennycook@intel.com>