Replace is_defined() with has_macro()
is_defined() confusingly returned a string instead of a bool,
and has_macro() is more aligned with the naming of get_macro().

This change also moves the handling of defined() into a single place,
making it clearer that defined() will return a "1" or "0" wrapped in a
NumericalConstant based on whether the macro exists.

Signed-off-by: John Pennycook <john.pennycook@intel.com>